cls/cas/cls_cas_internal: use DENC for refs_by_hash
In order to make the refs_t subtypes coexist with different encoding
schemes, I had to move things into the .cc file (probably better anyway)
and make them standalone classes (not subclasses).
