+++ src/cls/CMakeLists.txt
@@ -251,9 +251,18 @@ set(cls_lua_client_srcs
 add_library(cls_lua_client STATIC ${cls_lua_client_srcs})
 
 # cls_cas
+set(cls_cas_client_srcs
+  cas/cls_cas_client.cc)
+add_library(cls_cas_client STATIC ${cls_cas_client_srcs})
+
+set(cls_cas_internal_srcs
+  cas/cls_cas_internal.cc)
+add_library(cls_cas_internal STATIC ${cls_cas_internal_srcs})
+
 set(cls_cas_srcs 
   cas/cls_cas.cc)
 add_library(cls_cas SHARED ${cls_cas_srcs})
+target_link_libraries(cls_cas cls_cas_internal)
 set_target_properties(cls_cas PROPERTIES
   VERSION "1.0.0"
   SOVERSION "1"
@@ -261,9 +270,6 @@ set_target_properties(cls_cas PROPERTIES
   CXX_VISIBILITY_PRESET hidden)
 install(TARGETS cls_cas DESTINATION ${cls_dir})
 
-set(cls_cas_client_srcs
-  cas/cls_cas_client.cc)
-add_library(cls_cas_client STATIC ${cls_cas_client_srcs})
 
 #cls_queue
 set(cls_queue_srcs
