+++ src/test/librados/CMakeLists.txt
@@ -132,7 +132,7 @@ add_executable(ceph_test_rados_api_tier_pp
   tier_cxx.cc
   $<TARGET_OBJECTS:unit-main>)
 target_link_libraries(ceph_test_rados_api_tier_pp
-  librados global ${UNITTEST_LIBS} Boost::system radostest-cxx)
+  librados global ${UNITTEST_LIBS} Boost::system radostest-cxx cls_cas_internal)
 if(WITH_BOOST_CONTEXT)
   target_link_libraries(ceph_test_rados_api_tier_pp spawn)
 endif()
