+++ src/tools/CMakeLists.txt
@@ -118,7 +118,11 @@ install(TARGETS ceph-authtool DESTINATION bin)
 if(WITH_TESTS)
 set(ceph_dedup_tool_srcs ceph_dedup_tool.cc)
 add_executable(ceph-dedup-tool ${ceph_dedup_tool_srcs})
-target_link_libraries(ceph-dedup-tool librados global cls_cas_client)
+target_link_libraries(ceph-dedup-tool
+  librados
+  global
+  cls_cas_client
+  cls_cas_internal)
 install(TARGETS ceph-dedup-tool DESTINATION bin)
 endif(WITH_TESTS)
